requirements.txt
absl-py==0.8.1
astor==0.8.1
attrs==19.1.0
backcall==0.1.0
bleach==3.1.0
cachetools==3.1.1
cycler==0.10.0
decorator==4.4.0
defusedxml==0.6.0
dill==0.3.1.1
entrypoints==0.3
gast==0.2.2
google-auth==1.8.2
google-auth-oauthlib==0.4.1
google-pasta==0.1.8
gplearn==0.4.1
grpcio==1.25.0
h5py==2.10.0
iotop==0.6
ipykernel==5.1.2
ipython==7.8.0
ipython-genutils==0.2.0
jedi==0.15.1
Jinja2==2.10.1
joblib==0.14.1
json5==0.8.5
jsonschema==3.0.2
jupyter-client==5.3.2
jupyter-core==4.5.0
jupyterlab==1.1.3
jupyterlab-server==1.0.6
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
mistune==0.8.4
mock==3.0.5
nbconvert==5.6.0
nbformat==4.4.0
networkx==2.3
notebook==6.0.1
numpy==1.17.4
oauthlib==3.1.0
opt-einsum==3.1.0
ortools==7.3.7083
pandas==0.25.1
pandocfilters==1.4.2
parso==0.5.1
pexpect==4.7.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==2.0.9
protobuf==3.9.1
ptyprocess==0.6.0
PuLP==2.0
pyasn1==0.4.8
pyasn1-modules==0.2.7
Pygments==2.4.2
pyparsing==2.4.5
pyrsistent==0.15.4
python-dateutil==2.8.0
pytz==2019.2
PyYAML==5.2
pyzmq==18.1.0
requests-oauthlib==1.3.0
rsa==4.0
scikit-learn==0.22
scipy==1.4.0
seaborn==0.9.0
Send2Trash==1.5.0
six==1.13.0
sklearn==0.0
tb-nightly==1.14.0a20190301
tensorflow==1.13.1
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
tf-estimator-nightly==1.14.0.dev2019030115
torch==1.3.1
tornado==6.0.3
tqdm==4.35.0
traitlets==4.3.2
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.16.0
wrapt==1.11.2
xgboost==0.90
